Time: 31 ms (62.21%), Space: 21.5 MB (29.70%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -2,10 +2,10 @@ class Solution {
 public:
     int minEatingSpeed(vector<int>& piles, int h) {
         if(piles.size()>h) return -1; //if the answer is impossible.
-        long long l = 1, r = *max_element (piles.begin(), piles.end());
+        int l = 1, r = *max_element (piles.begin(), piles.end());
         int ans;
         while(l<=r){
-            long long m = (l+r)/2;
+            int m = (l+r)/2;
             long long totalH = timeTaken(piles, m);
             if(totalH<=h){
                 ans = m;
